Factor out DHCP client into a separate program
The main init process will (as a matter of policy) never interact directly
with IP sockets, to reduce its exposure to outside attacks. The plan
is for it to launch helper programs to do all interactions with the
outside world, and eventually to confine these helper programs with
AppArmor to reduce the possibility of them being used as attack vectors.

Here we create the first such applet, which obtains a DHCP lease from a
nearby server and notifies the main init process each time the lease
is renewed. It uses small msgpack-encoded messages written to its stdout
in order to notify the main init process, which is a pattern we expect
to repeat on other similar applets.

The defgrid-init program itself still remains a prototype stub, but now
it runs the separate dhcpclient applet rather than doing DHCP
configuration itself.