Overall structure of the boot process
The task of the boot process is to get the network stack configured and
discover this host's identity within the defgrid cluster. This is the
part that will be most variable between different hosting platforms, so
it's built in a generic way that supports multiple "flavors" that each
tailor the process to the particular platform.

For this initial commit, only the "dev" flavor is configured, which is
a set of stubs that can be used when running the program in a dev
environment without it interfering with the host system or requiring
root access.