Start of a real "main" implementation to boot the system
So far main has just been a prototyping space, but for the first time in
this commit it starts to do some real work for booting up a system,
building on the "booter" abstraction to support multiple different
platforms.

So far it just sets up the network and then hangs the machine, so still
not very useful but this is a reasonable checkpoint.