Generate RSA private host key on system boot
This key will be used for internal uses within the cluster, including as
the SSH host key and the key underlying TLS certs used for internal-only
services such as Consul. It is not intended for use on public services
intended for use by end-users.

The rationale for using a single host key for internal communication is
that during boot we have a limited entropy pool and so we can get more
"bang for our buck" this way, vs. trying to generate several keys that
may suffer from a lack of entropy.

The host key is managed by init and will be passed only to the specific
applications that need it: SSH server, Consul server, Vault server,
Nomad Server. This key will also later be used to generate a certificate
signing request to produce host TLS certs, e.g. using Vault's PKI backend.

The private key is not preserved across reboots because defgrid instances
are not expected to be rebooted in-place. Instead, an instance that is
no longer meeting expectations should be replaced by a new one.